Start adding spike notes
